add kafka test, make it integration test, fix mockBatch
